+++ pkg/cmd/tasks/execute/cli_inputs.go
@@ -3,57 +3,98 @@ package execute
 
 import (
 	"fmt"
+	"os"
 	"reflect"
 	"strconv"
 	"time"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/airplanedev/cli/pkg/api"
+	"github.com/airplanedev/cli/pkg/logger"
+	"github.com/airplanedev/cli/pkg/utils"
 	"github.com/pkg/errors"
 )
 
-func promptFromParam(param api.Parameter) (survey.Prompt, error) {
+// promptForParamValues attempts to prompt user for param values, setting them on `params`
+// If no TTY, errors unless there are no parameters
+// If TTY, prompts for parameters (if any) and asks user to confirm
+func promptForParamValues(client *api.Client, task api.Task, paramValues map[string]interface{}) error {
+	if !utils.CanPrompt() {
+		// Don't error if there are no params
+		if len(task.Parameters) == 0 {
+			return nil
+		}
+		// Otherwise, error since we have no params and no way to prompt for it
+		logger.Log("Parameters were not specified! Task has %d parameter(s):\n", len(task.Parameters))
+		for _, param := range task.Parameters {
+			var req string
+			if !param.Constraints.Optional {
+				req = "*"
+			}
+			logger.Log("  %s%s %s (%s)", param.Name, req, param.Type, param.Slug)
+			if param.Desc != "" {
+				logger.Log("    %s", param.Desc)
+			}
+		}
+		return errors.New("missing parameters")
+	}
+
+	logger.Log("You are about to run %s:", bold(task.Name))
+	logger.Log(gray(client.TaskURL(task.ID)))
+	logger.Log("")
+
+	for _, param := range task.Parameters {
+		prompt, err := promptForParam(param)
+		if err != nil {
+			return err
+		}
+		opts := []survey.AskOpt{
+			survey.WithStdio(os.Stdin, os.Stderr, os.Stderr),
+			survey.WithValidator(validateInput(param)),
+		}
+		if !param.Constraints.Optional {
+			opts = append(opts, survey.WithValidator(survey.Required))
+		}
+		var inputValue string
+		if err := survey.AskOne(prompt, &inputValue, opts...); err != nil {
+			return errors.Wrap(err, "asking prompt for param")
+		}
+
+		value, err := inputToAPIValue(param, inputValue)
+		if err != nil {
+			return errors.Wrap(err, "converting input to API value")
+		}
+		paramValues[param.Slug] = value
+	}
+	confirmed := false
+	if err := survey.AskOne(&survey.Confirm{
+		Message: "Execute?",
+		Default: true,
+	}, &confirmed); err != nil {
+		return errors.Wrap(err, "confirming")
+	}
+	if !confirmed {
+		return errors.New("user cancelled")
+	}
+	return nil
+}
+
+// promptForParam returns a survey.Prompt matching the param type
+func promptForParam(param api.Parameter) (survey.Prompt, error) {
 	message := fmt.Sprintf("%s (%s):", param.Name, param.Slug)
 	// TODO: support default values
 	switch param.Type {
-	case api.TypeString:
-		return &survey.Input{
-			Message: message,
-			Help:    param.Desc,
-		}, nil
 	case api.TypeBoolean:
 		return &survey.Select{
 			Message: message,
 			Help:    param.Desc,
 			Options: []string{"Yes", "No"},
 		}, nil
-	case api.TypeUpload:
-		return &survey.Input{
-			Message: message,
-			Help:    param.Desc,
-		}, nil
-	case api.TypeInteger:
-		return &survey.Input{
-			Message: message,
-			Help:    param.Desc,
-		}, nil
-	case api.TypeFloat:
-		return &survey.Input{
-			Message: message,
-			Help:    param.Desc,
-		}, nil
-	case api.TypeDate:
-		return &survey.Input{
-			Message: message,
-			Help:    param.Desc,
-		}, nil
-	case api.TypeDatetime:
-		return &survey.Input{
-			Message: message,
-			Help:    param.Desc,
-		}, nil
 	default:
-		return nil, errors.Errorf("unexpected param type %s", param.Type)
+		return &survey.Input{
+			Message: message,
+			Help:    param.Desc,
+		}, nil
 	}
 }
